Fetching the repository to main (#1)
* LeetCode submission problem (#31)

* Delete LeetCode/Math directory

* Add files via upload

* Time unit change and new CP Problem

Time unit change to (ms)

* Add files via upload

* Add LeetCode submission problem #4

* Add LeetCode submission solution

* Update README.md

Add LeetCode Submission


* refactor: update documentation and solutions directories (#32)

* docs: add reference to each category README file

* feat: HackerRank Solution - Interview Preparation Kit - Sorting

* docs: update README.md

* feat: HackerRank Solution - Interview Preparation Kit - Arrays

* refactor: separate solutions based on the problem difficulty

* docs: add problems category tags
