+++ LeetCode/Easy/README.md
@@ -1,0 +1,9 @@
+# CP Solution for Leet Code Easy Problems
+
+|**No**| **Problems**      | **Language Used** | **Time(ms)** | **Memory(MB)** | **Problem Tags** |
+| ---- | ----------------- |:--------:| --------:| ----------:| ---------- |
+| 1 | [Decoded XORed Array](./Decoded_XORed_Array.cpp) | C++ | 36	| 24.7 | ```Bit Manipulation``` |
+| 2 | [Number of Steps to Reduce a Number to Zero](./NumberOfStepsToReduceANumberToZero.cpp) | C++ | 4 | 5.9 | ```Bit Manipulation``` |
+| 3 | [Number of Good Pairs](./NumberOfGoodPairs.py) | Python | 28 | 14.2 | ```Math``` |
+| 4 | [Concatenation of Array](./ConcatenationArray.cpp)| C++ | 8 | 12.5 | ```Array``` |
+| 5 | [Two Sum](./TwoSum.cpp) | C++ | 16 | 11.9 | ```Array``` |
